ci: fix clippy warnings - remove unnecessary casts and fix formatting
- Removed unnecessary casts in terminal_engine (grid, mod, parser, session)
- Fixed redundant field names in struct initialization
- Removed unnecessary parentheses and fixed formatting
- All 11 clippy errors resolved, clippy and rustfmt checks now pass

+++ src-tauri/src/terminal_engine/grid.rs
@@ -37,7 +37,7 @@ impl GridBuffer {
             self.cursor.row += 1;
             if self.cursor.row >= self.rows {
                 self.scroll_up(1);
-                self.cursor.row = (self.rows - 1) as u16;
+                self.cursor.row = self.rows - 1;
             }
         }
     }
@@ -92,7 +92,7 @@ impl GridBuffer {
                     self.cursor.row += 1;
                     if self.cursor.row >= self.rows {
                         self.scroll_up(1);
-                        self.cursor.row = (self.rows - 1) as u16;
+                        self.cursor.row = self.rows - 1;
                     }
                 }
                 '\t' => {
